+++ CommonConcepts/DslScripts/SQL/Log.Created default.MsSql.sql
@@ -1,1 +1,1 @@
-getdate()
+SYSDATETIME()