+++ CommonConcepts/Plugins/Rhetos.Dom.DefaultConcepts/DataStructure/DomInitializationCodeGenerator.cs
@@ -42,11 +42,13 @@ namespace Rhetos.Dom.DefaultConcepts
 
         private readonly RhetosBuildEnvironment _buildEnvironment;
         private readonly CommonConceptsOptions _commonConceptsOptions;
+        private readonly CommonConceptsDatabaseSettings _databaseSettings;
 
-        public DomInitializationCodeGenerator(RhetosBuildEnvironment buildEnvironment, CommonConceptsOptions commonConceptsOptions)
+        public DomInitializationCodeGenerator(RhetosBuildEnvironment buildEnvironment, CommonConceptsOptions commonConceptsOptions, CommonConceptsDatabaseSettings databaseSettings)
         {
             _buildEnvironment = buildEnvironment;
             _commonConceptsOptions = commonConceptsOptions;
+            _databaseSettings = databaseSettings;
         }
 
         public static readonly string StandardNamespacesSnippet =
@@ -85,6 +87,7 @@ namespace Rhetos.Dom.DefaultConcepts
             codeBuilder.AddReferencesFromDependency(typeof(System.Data.Entity.Core.Objects.ObjectStateEntry));
             codeBuilder.AddReferencesFromDependency(typeof(Rhetos.Persistence.IPersistenceCache));
             codeBuilder.AddReferencesFromDependency(typeof(Rhetos.Persistence.IPersistenceTransaction));
+            codeBuilder.AddReferencesFromDependency(typeof(Rhetos.Dom.DefaultConcepts.CommonConceptsDatabaseSettings));
             codeBuilder.AddReferencesFromDependency(typeof(ApplyFiltersOnClientRead));
             codeBuilder.AddReferencesFromDependency(typeof(ICommandInfo)); // Used from ApplyFiltersOnClientRead.
         }
@@ -235,7 +238,7 @@ $@"namespace Common
 
         {ModuleCodeGenerator.CommonDomRepositoryMembersTag}
     }}
-
+    
     public static class Infrastructure
     {{
         public static readonly RegisteredInterfaceImplementations RegisteredInterfaceImplementations = new RegisteredInterfaceImplementations
@@ -321,6 +324,7 @@ $@"namespace Common
     }}
 
     [System.ComponentModel.Composition.Export(typeof(Autofac.Module))]
+    [System.ComponentModel.Composition.ExportMetadata(Rhetos.Extensibility.MefProvider.DependsOn, typeof(Rhetos.Dom.DefaultConcepts.AutofacModuleConfiguration))] // Overrides some registrations from that class.
     public class AutofacModuleConfiguration : Autofac.Module
     {{
         protected override void Load(Autofac.ContainerBuilder builder)
@@ -337,6 +341,13 @@ $@"namespace Common
             builder.RegisterType<ExecutionContext>().InstancePerLifetimeScope();
             builder.RegisterInstance(Infrastructure.RegisteredInterfaceImplementations).ExternallyOwned();
             builder.RegisterInstance(Infrastructure.ApplyFiltersOnClientRead).ExternallyOwned();
+            builder.RegisterInstance(new Rhetos.Dom.DefaultConcepts.CommonConceptsDatabaseSettings
+            {{
+                UseLegacyMsSqlDateTime = {_databaseSettings.UseLegacyMsSqlDateTime.ToString().ToLowerInvariant()},
+                DateTimePrecision = {_databaseSettings.DateTimePrecision},
+            }});
+            builder.Register<CommonConceptsOptions>(context => throw new NotImplementedException($""{{nameof(CommonConceptsOptions)}} is a build-time configuration, not available at run-time.""));
+            
             {ModuleCodeGenerator.CommonAutofacConfigurationMembersTag}
 
             base.Load(builder);
@@ -398,7 +409,6 @@ $@"namespace Common
         where TEntity : class, IEntity
         where TQueryableEntity : class, IEntity, TEntity, IQueryableEntity<TEntity>
     {{
-        [Obsolete(""Use Load(ids) or Query(ids) method."")]
         public TEntity[] Load(IEnumerable<Guid> ids)
         {{
             if (!(ids is System.Collections.IList))