+++ CommonConcepts/Plugins/Rhetos.Dsl.DefaultConcepts/Advanced/EntityHistoryInfo.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
     Copyright (C) 2014 Omega software d.o.o.
 
     This file is part of Rhetos.
@@ -17,14 +17,12 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
+using Rhetos.Compiler;
+using Rhetos.Dom.DefaultConcepts;
+using Rhetos.Utilities;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.ComponentModel.Composition;
-using Rhetos.Utilities;
-using Rhetos.Compiler;
-using System.Globalization;
 
 namespace Rhetos.Dsl.DefaultConcepts
 {
@@ -34,58 +32,73 @@ namespace Rhetos.Dsl.DefaultConcepts
     /// </summary>
     [Export(typeof(IConceptInfo))]
     [ConceptKeyword("History")]
-    public class EntityHistoryInfo : IMacroConcept, IAlternativeInitializationConcept
+    public class EntityHistoryInfo : IAlternativeInitializationConcept
     {
         [ConceptKey]
         public EntityInfo Entity { get; set; }
 
         public EntityInfo Dependency_ChangesEntity { get; set; }
-        public SqlQueryableInfo Dependency_HistorySqlQueryable { get; set; }
-        public SqlFunctionInfo Dependency_AtTimeSqlFunction { get; set; }
-        public WriteInfo Dependency_Write { get; set; }
-        
-        public static readonly CsTag<EntityHistoryInfo> ClonePropertiesTag = "ClonePropertiesRewrite";
 
-        public IEnumerable<string> DeclareNonparsableProperties()
-        {
-            return new string[] { "Dependency_ChangesEntity", "Dependency_HistorySqlQueryable", "Dependency_AtTimeSqlFunction", "Dependency_Write" };
-        }
+        public IEnumerable<string> DeclareNonparsableProperties() => new[] { "Dependency_ChangesEntity" };
 
         public void InitializeNonparsableProperties(out IEnumerable<IConceptInfo> createdConcepts)
         {
             Dependency_ChangesEntity = new EntityInfo { Module = Entity.Module, Name = Entity.Name + "_Changes" };
-            Dependency_HistorySqlQueryable = new SqlQueryableInfo { Module = Entity.Module, Name = Entity.Name + "_History", SqlSource = HistorySqlSnippet() };
-            Dependency_AtTimeSqlFunction = new SqlFunctionInfo { Module = Entity.Module, Name = Entity.Name + "_AtTime", Arguments = "@ContextTime DATETIME", Source = AtTimeSqlSnippet() };
-            Dependency_Write = new WriteInfo {
-                    DataStructure = Dependency_HistorySqlQueryable,
-                    SaveImplementation = HistorySaveFunction()
-                };
 
-            createdConcepts = new IConceptInfo[] { Dependency_ChangesEntity, Dependency_HistorySqlQueryable, Dependency_AtTimeSqlFunction, Dependency_Write };        
+            createdConcepts = new IConceptInfo[] { Dependency_ChangesEntity };
+        }
+    }
+
+    [Export(typeof(IConceptMacro))]
+    public class EntityHistoryMacro : IConceptMacro<EntityHistoryInfo>
+    {
+        public static readonly CsTag<EntityHistoryInfo> ClonePropertiesTag = "ClonePropertiesRewrite";
+        public static readonly SqlTag<EntityHistoryInfo> SelectHistoryPropertiesTag = "SelectHistoryProperties";
+        public static readonly SqlTag<EntityHistoryInfo> SelectEntityPropertiesTag = "SelectEntityProperties";
+
+        private readonly string _dateTimeSqlColumnType;
+
+        public EntityHistoryMacro(CommonConceptsDatabaseSettings databaseSettings)
+        {
+            // TODO: we inject DatabaseSetttings here, but correct solution would be to inject ConceptMetadata and use ConceptMetadata.GetColumnType to fetch SQL type of the column
+            // we can't do that because of circular reference problem between projects and IDatabaseColumnType is not available
+            // needs to be refactored after we rearrange/merge projects and remove circular dependency in question
+            _dateTimeSqlColumnType = databaseSettings.UseLegacyMsSqlDateTime
+                ? "DATETIME"
+                : $"DATETIME2({databaseSettings.DateTimePrecision})";
         }
 
-        public IEnumerable<IConceptInfo> CreateNewConcepts(IEnumerable<IConceptInfo> existingConcepts)
+        public IEnumerable<IConceptInfo> CreateNewConcepts(EntityHistoryInfo conceptInfo, IDslModel existingConcepts)
         {
             var newConcepts = new List<IConceptInfo>();
 
+            var atTimeSqlFunction = new SqlFunctionInfo { Module = conceptInfo.Entity.Module, Name = conceptInfo.Entity.Name + "_AtTime", Arguments = $"@ContextTime {_dateTimeSqlColumnType}", Source = AtTimeSqlSnippet(conceptInfo) };
+            newConcepts.Add(atTimeSqlFunction);
+
+            var historySqlQueryable = new SqlQueryableInfo { Module = conceptInfo.Entity.Module, Name = conceptInfo.Entity.Name + "_History", SqlSource = HistorySqlSnippet(conceptInfo) };
+            newConcepts.Add(historySqlQueryable);
+
+            var write = new WriteInfo { DataStructure = historySqlQueryable, SaveImplementation = HistorySaveFunction(conceptInfo) };
+            newConcepts.Add(write);
+
             // Expand the base entity:
-            var activeSinceProperty = new DateTimePropertyInfo { DataStructure = Entity, Name = "ActiveSince" }; // TODO: SystemRequired, Default 1.1.1900.
+            var activeSinceProperty = new DateTimePropertyInfo { DataStructure = conceptInfo.Entity, Name = "ActiveSince" }; // TODO: SystemRequired, Default 1.1.1900., after implementing optimization for multiple simple property validations.
             var activeSinceHistory = new EntityHistoryPropertyInfo { Property = activeSinceProperty };
             newConcepts.AddRange(new IConceptInfo[] { activeSinceProperty, activeSinceHistory });
 
             // InvalidData for base entity: it is not allowed to save with ActiveSince older than last one used in History
             var denyFilter = new ComposableFilterByInfo {
                 Parameter = "Common.OlderThanHistoryEntries",
-                Source = Entity,
+                Source = conceptInfo.Entity,
                 Expression = String.Format(
                     @"(items, repository, parameter) => items.Where(item => 
-                                repository.{0}.{1}_Changes.Subquery.Where(his => his.ActiveSince >= item.ActiveSince && his.Entity == item).Count() > 0)", 
-                                Entity.Module.Name, 
-                                Entity.Name) 
+                                repository.{0}.{1}_Changes.Subquery.Where(his => his.ActiveSince >= item.ActiveSince && his.Entity == item).Count() > 0)",
+                                conceptInfo.Entity.Module.Name,
+                                conceptInfo.Entity.Name) 
             };
             var invalidDataValidation = new InvalidDataInfo {
                 FilterType = "Common.OlderThanHistoryEntries", 
-                Source = Entity, 
+                Source = conceptInfo.Entity, 
                 ErrorMessage = "ActiveSince is not allowed to be older than last entry in history."
             };
             newConcepts.AddRange(new IConceptInfo[]
@@ -97,28 +110,28 @@ namespace Rhetos.Dsl.DefaultConcepts
                 });
 
             // Create a new entity for history data:
-            var currentProperty = new ReferencePropertyInfo { DataStructure = Dependency_ChangesEntity, Name = "Entity", Referenced = Entity };
-            var historyActiveSinceProperty = new DateTimePropertyInfo { DataStructure = Dependency_ChangesEntity, Name = activeSinceProperty.Name };
+            var currentProperty = new ReferencePropertyInfo { DataStructure = conceptInfo.Dependency_ChangesEntity, Name = "Entity", Referenced = conceptInfo.Entity };
+            var historyActiveSinceProperty = new DateTimePropertyInfo { DataStructure = conceptInfo.Dependency_ChangesEntity, Name = activeSinceProperty.Name };
             newConcepts.AddRange(new IConceptInfo[] {
                 currentProperty,
                 new ReferenceDetailInfo { Reference = currentProperty },
-                new RequiredPropertyInfo { Property = currentProperty }, // TODO: SystemRequired
-                new PropertyFromInfo { Destination = Dependency_ChangesEntity, Source = activeSinceProperty },
+                new RequiredPropertyInfo { Property = currentProperty }, // TODO: SystemRequired, after implementing optimization for multiple simple property validations.
+                new PropertyFromInfo { Destination = conceptInfo.Dependency_ChangesEntity, Source = activeSinceProperty },
                 historyActiveSinceProperty,
-                new UniqueMultiplePropertiesInfo { DataStructure = Dependency_ChangesEntity, PropertyNames = $"{currentProperty.Name} {historyActiveSinceProperty.Name}" }
+                new UniqueMultiplePropertiesInfo { DataStructure = conceptInfo.Dependency_ChangesEntity, PropertyNames = $"{currentProperty.Name} {historyActiveSinceProperty.Name}" }
             });
 
             // InvalidData for history entity: it is not allowed to save with ActiveSince newer than current entity
             var denyFilterHistory = new ComposableFilterByInfo
             {
                 Parameter = "Common.NewerThanCurrentEntry",
-                Source = Dependency_ChangesEntity,
+                Source = conceptInfo.Dependency_ChangesEntity,
                 Expression = @"(items, repository, parameter) => items.Where(item => item.ActiveSince > item.Entity.ActiveSince)"
             };
             var invalidDataValidationHistory = new InvalidDataInfo
             {
                 FilterType = "Common.NewerThanCurrentEntry",
-                Source = Dependency_ChangesEntity,
+                Source = conceptInfo.Dependency_ChangesEntity,
                 ErrorMessage = "ActiveSince of history entry is not allowed to be newer than current entry."
             };
             newConcepts.AddRange(new IConceptInfo[]
@@ -130,31 +143,31 @@ namespace Rhetos.Dsl.DefaultConcepts
                 });
 
             // Create ActiveUntil SqlQueryable:
-            var activeUntilSqlQueryable = new SqlQueryableInfo { Module = Entity.Module, Name = Entity.Name + "_ChangesActiveUntil", SqlSource = ActiveUntilSqlSnippet() };
+            var activeUntilSqlQueryable = new SqlQueryableInfo { Module = conceptInfo.Entity.Module, Name = conceptInfo.Entity.Name + "_ChangesActiveUntil", SqlSource = ActiveUntilSqlSnippet(conceptInfo) };
             newConcepts.AddRange(new IConceptInfo[] {
                 activeUntilSqlQueryable,
                 new DateTimePropertyInfo { DataStructure = activeUntilSqlQueryable, Name = "ActiveUntil" },
-                new SqlDependsOnDataStructureInfo { Dependent = activeUntilSqlQueryable, DependsOn = Dependency_ChangesEntity },
-                new SqlDependsOnDataStructureInfo { Dependent = activeUntilSqlQueryable, DependsOn = Entity },
-                new DataStructureExtendsInfo { Base = Dependency_ChangesEntity, Extension = activeUntilSqlQueryable }
+                new SqlDependsOnDataStructureInfo { Dependent = activeUntilSqlQueryable, DependsOn = conceptInfo.Dependency_ChangesEntity },
+                new SqlDependsOnDataStructureInfo { Dependent = activeUntilSqlQueryable, DependsOn = conceptInfo.Entity },
+                new DataStructureExtendsInfo { Base = conceptInfo.Dependency_ChangesEntity, Extension = activeUntilSqlQueryable }
             });
 
             // Configure History SqlQueryable:
             newConcepts.AddRange(new IConceptInfo[] {
-                new SqlDependsOnDataStructureInfo { Dependent = Dependency_HistorySqlQueryable, DependsOn = Entity },
-                new SqlDependsOnDataStructureInfo { Dependent = Dependency_HistorySqlQueryable, DependsOn = Dependency_ChangesEntity },
-                new SqlDependsOnDataStructureInfo { Dependent = Dependency_HistorySqlQueryable, DependsOn = activeUntilSqlQueryable },
-                new DateTimePropertyInfo { DataStructure = Dependency_HistorySqlQueryable, Name = "ActiveUntil" },
-                new AllPropertiesFromInfo { Source = Dependency_ChangesEntity, Destination = Dependency_HistorySqlQueryable }
+                new SqlDependsOnDataStructureInfo { Dependent = historySqlQueryable, DependsOn = conceptInfo.Entity },
+                new SqlDependsOnDataStructureInfo { Dependent = historySqlQueryable, DependsOn = conceptInfo.Dependency_ChangesEntity },
+                new SqlDependsOnDataStructureInfo { Dependent = historySqlQueryable, DependsOn = activeUntilSqlQueryable },
+                new DateTimePropertyInfo { DataStructure = historySqlQueryable, Name = "ActiveUntil" },
+                new AllPropertiesFromInfo { Source = conceptInfo.Dependency_ChangesEntity, Destination = historySqlQueryable }
             });
 
             // Configure AtTime SqlFunction:
-            newConcepts.Add(new SqlDependsOnDataStructureInfo { Dependent = Dependency_AtTimeSqlFunction, DependsOn = Dependency_HistorySqlQueryable });
+            newConcepts.Add(new SqlDependsOnDataStructureInfo { Dependent = atTimeSqlFunction, DependsOn = historySqlQueryable });
 
             return newConcepts;
         }
 
-        private string ActiveUntilSqlSnippet()
+        private string ActiveUntilSqlSnippet(EntityHistoryInfo conceptInfo)
         {
             return string.Format(
                 @"SELECT
@@ -166,18 +179,18 @@ namespace Rhetos.Dsl.DefaultConcepts
 				                newerVersion.ActiveSince > history.ActiveSince
 	                INNER JOIN {0}.{1} currentItem ON currentItem.ID = history.EntityID
                 GROUP BY history.ID",
-                    SqlUtility.Identifier(Entity.Module.Name),
-                    SqlUtility.Identifier(Entity.Name),
-                    SqlUtility.Identifier(Dependency_ChangesEntity.Name));
+                    SqlUtility.Identifier(conceptInfo.Entity.Module.Name),
+                    SqlUtility.Identifier(conceptInfo.Entity.Name),
+                    SqlUtility.Identifier(conceptInfo.Dependency_ChangesEntity.Name));
         }
 
-        private string HistorySqlSnippet()
+        private string HistorySqlSnippet(EntityHistoryInfo conceptInfo)
         {
             return string.Format(
                 @"SELECT
                     ID = entity.ID,
                     EntityID = entity.ID,
-                    ActiveUntil = CAST(NULL AS DateTime){5}
+                    ActiveUntil = CAST(NULL AS {6}){5}
                 FROM
                     {0}.{1} entity
 
@@ -190,15 +203,16 @@ namespace Rhetos.Dsl.DefaultConcepts
                 FROM
                     {0}.{2} history
                     LEFT JOIN {0}.{3} au ON au.ID = history.ID",
-                SqlUtility.Identifier(Entity.Module.Name),
-                SqlUtility.Identifier(Entity.Name),
-                SqlUtility.Identifier(Dependency_ChangesEntity.Name),
-                SqlUtility.Identifier(Entity.Name + "_ChangesActiveUntil"),
-                SelectHistoryPropertiesTag.Evaluate(this),
-                SelectEntityPropertiesTag.Evaluate(this));
+                SqlUtility.Identifier(conceptInfo.Entity.Module.Name),
+                SqlUtility.Identifier(conceptInfo.Entity.Name),
+                SqlUtility.Identifier(conceptInfo.Dependency_ChangesEntity.Name),
+                SqlUtility.Identifier(conceptInfo.Entity.Name + "_ChangesActiveUntil"),
+                SelectHistoryPropertiesTag.Evaluate(conceptInfo),
+                SelectEntityPropertiesTag.Evaluate(conceptInfo),
+                _dateTimeSqlColumnType);
         }
 
-        private string AtTimeSqlSnippet()
+        private string AtTimeSqlSnippet(EntityHistoryInfo conceptInfo)
         {
             return string.Format(
                 @"RETURNS TABLE
@@ -217,15 +231,12 @@ namespace Rhetos.Dsl.DefaultConcepts
                             WHERE ActiveSince <= @ContextTime
                             GROUP BY EntityID
                         ) last ON last.EntityID = history.EntityID AND last.Max_ActiveSince = history.ActiveSince",
-                    SqlUtility.Identifier(Entity.Module.Name),
-                    SqlUtility.Identifier(Entity.Name + "_History"),
-                    SelectHistoryPropertiesTag.Evaluate(this));
+                    SqlUtility.Identifier(conceptInfo.Entity.Module.Name),
+                    SqlUtility.Identifier(conceptInfo.Entity.Name + "_History"),
+                    SelectHistoryPropertiesTag.Evaluate(conceptInfo));
         }
 
-        public static readonly SqlTag<EntityHistoryInfo> SelectHistoryPropertiesTag = "SelectHistoryProperties";
-        public static readonly SqlTag<EntityHistoryInfo> SelectEntityPropertiesTag = "SelectEntityProperties";
-
-        private string HistorySaveFunction()
+        private string HistorySaveFunction(EntityHistoryInfo conceptInfo)
         {
             return String.Format(@"var updateEnt = new List<{0}_History>();
             var deletedEnt = new List<{0}_History>();
@@ -315,10 +326,10 @@ namespace Rhetos.Dsl.DefaultConcepts
             _domRepository.{1}.{0}.Save(null, updateCurrentAndAddHistory, deletedEnt.Select(de => new {1}.{0} {{ ID = de.EntityID.Value }}));
             _domRepository.{1}.{0}.Save(null, updateCurrentItemsOnly, null);
 
-            ",
-             Entity.Name,
-             Entity.Module.Name,
-             ClonePropertiesTag.Evaluate(this));
+            ", 
+             conceptInfo.Entity.Name,
+             conceptInfo.Entity.Module.Name,
+             ClonePropertiesTag.Evaluate(conceptInfo));
         }
     }
 }